Restore across service plans (#140)
* Restore across service plans

With this change Service Fabrik should be able to restore the backup of
one plan across plans of the same instance. Only backups taken from
plans specified in the restore_predecessor section of the current plan
can be restored. If the restore_predecessor section is absent in the
plan update_predecessor is checked.

* Broker script changes for working with bosh-lite: Added `--post302`
  which is necessary to query broker when the url has redirection.
* Added restore_predecessors check for update operation.
* Removed plan_id from backup and restore file name.
* Unit and acceptance tests.

* Use more informative message

* Better variable name